Implemented DecisionTree.py
I made a few minor grammatical fixes/tweaks to the ProjectPlan.docx and also switched it to README.md, which should be easier to view on GitHub, since GitHub does not play nicely with MS Office documents.

I also implemented the DecisionTree.py script, which just eats a training dataset and spits out the performance metrics for the decision tree across an array of different depths.