Regenerated Figures and Updated Draft
I updated the figures to use the proper names of things, and I made a few slight adjustments to the report draft.

It's also time to play my favorite game: how to anger every statistician on the planet. Let's see how many cardinal sins I committed with the Hotelling's T-test.